Move fulfillment related code to src folder
